add first note combination logic

+++ src/DesafioItau.Application/UseCases/V1/CashWithdrawal/GetNotesCombination/GetNotesCombinationRequest.cs
@@ -2,5 +2,5 @@
 
 public class GetNotesCombinationRequest
 {
-    public decimal? Amount { get; set; }
+    public int? Amount { get; set; }
 }
